Scope of the classes changed from private to protected and removed static flag
