+++ src/main/java/tech/artisanhub/ShapeletTrainer/OrderLineObj.java
@@ -1,0 +1,12 @@
+package tech.artisanhub.ShapeletTrainer;
+
+public class OrderLineObj {
+
+    protected double distance;
+    protected double classVal;
+
+    protected OrderLineObj(double distance, double classVal) {
+        this.distance = distance;
+        this.classVal = classVal;
+    }
+}
