+++ src/main/java/tech/artisanhub/ShapeletTrainer/ShapeletFilter.java
@@ -127,7 +127,7 @@ public class ShapeletFilter {
      * @return the transformed Instances in terms of the distance from each shapelet
      * @throws Exception - if the number of shapelets or the length parameters specified are incorrect
      */
-    public Instances process(Instances data) throws Exception {
+    public ArrayList<Shapelet> process(Instances data) throws Exception {
         if (this.numShapelets < 1) {
             throw new Exception(
                     "Number of shapelets initialised incorrectly - please select value of k (Usage: setNumberOfShapelets");
@@ -153,20 +153,20 @@ public class ShapeletFilter {
             System.out.println(shapelets.size() + " Shapelets have been generated");
         }
 
-        Instances output = determineOutputFormat(data);
-
-        // for each data, get distance to each shapelet and create new instance
-        for (int i = 0; i < data.numInstances(); i++) { // for each data
-            Instance toAdd = new SparseInstance(this.shapelets.size() + 1);
-            int shapeletNum = 0;
-            for (Shapelet s : this.shapelets) {
-                double dist = subsequenceDistance(s.content, data.instance(i));
-                toAdd.setValue(shapeletNum++, dist);
-            }
-            toAdd.setValue(this.shapelets.size(), data.instance(i).classValue());
-            output.add(toAdd);
-        }
-        return output;
+//        Instances output = determineOutputFormat(data);
+//
+//        // for each data, get distance to each shapelet and create new instance
+//        for (int i = 0; i < data.numInstances(); i++) { // for each data
+//            Instance toAdd = new SparseInstance(this.shapelets.size() + 1);
+//            int shapeletNum = 0;
+//            for (Shapelet s : this.shapelets) {
+//                double dist = subsequenceDistance(s.content, data.instance(i));
+//                toAdd.setValue(shapeletNum++, dist);
+//            }
+//            toAdd.setValue(this.shapelets.size(), data.instance(i).classValue());
+//            output.add(toAdd);
+ //       }
+        return shapelets;
     }
 
     public void setLogOutputFile(String fileName) {
@@ -423,37 +423,6 @@ public class ShapeletFilter {
         }
     }
 
-    private static double entropy(TreeMap<Double, Integer> classDistributions) {
-        if (classDistributions.size() == 1) {
-            return 0;
-        }
-
-        double thisPart;
-        double toAdd;
-        int total = 0;
-        for (Double d : classDistributions.keySet()) {
-            total += classDistributions.get(d);
-        }
-        // to avoid NaN calculations, the individual parts of the entropy are calculated and summed.
-        // i.e. if there is 0 of a class, then that part would calculate as NaN, but this can be caught
-        // and
-        // set to 0.
-        ArrayList<Double> entropyParts = new ArrayList<Double>();
-        for (Double d : classDistributions.keySet()) {
-            thisPart = (double) classDistributions.get(d) / total;
-            toAdd = -thisPart * Math.log10(thisPart) / Math.log10(2);
-            if (Double.isNaN(toAdd))
-                toAdd = 0;
-            entropyParts.add(toAdd);
-        }
-
-        double entropy = 0;
-        for (int i = 0; i < entropyParts.size(); i++) {
-            entropy += entropyParts.get(i);
-        }
-        return entropy;
-    }
-
     /**
      *
      * @param candidate
@@ -569,201 +538,7 @@ public class ShapeletFilter {
         return false;
     }
 
-    private static class Shapelet implements Comparable<Shapelet> {
-        private double[] content;
-        private int seriesId;
-        private int startPos;
-        private double splitThreshold;
-        private double informationGain;
-        private double separationGap;
 
-        private Shapelet(double[] content, int seriesId, int startPos) {
-            this.content = content;
-            this.seriesId = seriesId;
-            this.startPos = startPos;
-        }
-
-        // TEMPORARY - for testing
-        private Shapelet(double[] content, int seriesId, int startPos, double splitThreshold,
-                         double gain, double gap) {
-            this.content = content;
-            this.seriesId = seriesId;
-            this.startPos = startPos;
-            this.splitThreshold = splitThreshold;
-            this.informationGain = gain;
-            this.separationGap = gap;
-        }
-
-        // TEMP - used when processing has been carried out in initial stage, then the shapelets read in
-        // via csv later
-        private Shapelet(double[] content) {
-            this.content = content;
-        }
-
-    /*
-     * note: we calculate the threshold as this is used for finding the best split point in the data
-     * however, as this implementation of shapelets is as a filter, we do not actually use the
-     * threshold in the transformation.
-     */
-
-        private void calcInfoGainAndThreshold(ArrayList<OrderLineObj> orderline,
-                                              TreeMap<Double, Integer> classDistribution) {
-            // for each split point, starting between 0 and 1, ending between end-1 and end
-            // addition: track the last threshold that was used, don't bother if it's the same as the last
-            // one
-            double lastDist = orderline.get(0).distance; // must be initialised as not visited(no point
-            // breaking before any data!)
-            double thisDist = -1;
-
-            double bsfGain = -1;
-            double threshold = -1;
-
-            // check that there is actually a split point
-            // for example, if all
-
-            for (int i = 1; i < orderline.size(); i++) {
-                thisDist = orderline.get(i).distance;
-                if (i == 1 || thisDist != lastDist) { // check that threshold has moved(no point in sampling
-                    // identical thresholds)- special case - if 0 and 1
-                    // are the same dist
-
-                    // count class instances below and above threshold
-                    TreeMap<Double, Integer> lessClasses = new TreeMap<Double, Integer>();
-                    TreeMap<Double, Integer> greaterClasses = new TreeMap<Double, Integer>();
-
-                    for (double j : classDistribution.keySet()) {
-                        lessClasses.put(j, 0);
-                        greaterClasses.put(j, 0);
-                    }
-
-                    int sumOfLessClasses = 0;
-                    int sumOfGreaterClasses = 0;
-
-                    // visit those below threshold
-                    for (int j = 0; j < i; j++) {
-                        double thisClassVal = orderline.get(j).classVal;
-                        int storedTotal = lessClasses.get(thisClassVal);
-                        storedTotal++;
-                        lessClasses.put(thisClassVal, storedTotal);
-                        sumOfLessClasses++;
-                    }
-
-                    // visit those above threshold
-                    for (int j = i; j < orderline.size(); j++) {
-                        double thisClassVal = orderline.get(j).classVal;
-                        int storedTotal = greaterClasses.get(thisClassVal);
-                        storedTotal++;
-                        greaterClasses.put(thisClassVal, storedTotal);
-                        sumOfGreaterClasses++;
-                    }
-
-                    int sumOfAllClasses = sumOfLessClasses + sumOfGreaterClasses;
-
-                    double parentEntropy = entropy(classDistribution);
-
-                    // calculate the info gain below the threshold
-                    double lessFrac = (double) sumOfLessClasses / sumOfAllClasses;
-                    double entropyLess = entropy(lessClasses);
-                    // calculate the info gain above the threshold
-                    double greaterFrac = (double) sumOfGreaterClasses / sumOfAllClasses;
-                    double entropyGreater = entropy(greaterClasses);
-
-                    double gain = parentEntropy - lessFrac * entropyLess - greaterFrac * entropyGreater;
-                    if (gain > bsfGain) {
-                        bsfGain = gain;
-                        threshold = (thisDist - lastDist) / 2 + lastDist;
-                    }
-                }
-                lastDist = thisDist;
-            }
-            if (bsfGain >= 0) {
-                this.informationGain = bsfGain;
-                this.splitThreshold = threshold;
-                this.separationGap = calculateSeparationGap(orderline, threshold);
-            }
-        }
-
-        private double calculateSeparationGap(ArrayList<OrderLineObj> orderline,
-                                              double distanceThreshold) {
-
-            double sumLeft = 0;
-            double leftSize = 0;
-            double sumRight = 0;
-            double rightSize = 0;
-
-            for (int i = 0; i < orderline.size(); i++) {
-                if (orderline.get(i).distance < distanceThreshold) {
-                    sumLeft += orderline.get(i).distance;
-                    leftSize++;
-                }
-                else {
-                    sumRight += orderline.get(i).distance;
-                    rightSize++;
-                }
-            }
-
-            double thisSeparationGap = 1 / rightSize * sumRight - 1 / leftSize * sumLeft;
-
-            if (rightSize == 0 || leftSize == 0) {
-                return -1; // obviously there was no seperation, which is likely to be very rare but i still
-                // caused it!
-            } // e.g if all data starts with 0, first shapelet length =1, there will be no seperation as
-            // all time series are same dist
-            // equally true if all data contains the shapelet candidate, which is a more realistic example
-
-            return thisSeparationGap;
-        }
-
-        // comparison 1: to determine order of shapelets in terms of info gain, then separation gap,
-        // then shortness
-        public int compareTo(Shapelet shapelet) {
-            final int BEFORE = -1;
-            final int EQUAL = 0;
-            final int AFTER = 1;
-
-            if (this.informationGain != shapelet.informationGain) {
-                if (this.informationGain > shapelet.informationGain) {
-                    return BEFORE;
-                }
-                else {
-                    return AFTER;
-                }
-            }
-            else {
-                if (this.separationGap != shapelet.separationGap) {
-                    if (this.separationGap > shapelet.separationGap) {
-                        return BEFORE;
-                    }
-                    else {
-                        return AFTER;
-                    }
-                }
-                else if (this.content.length != shapelet.content.length) {
-                    if (this.content.length < shapelet.content.length) {
-                        return BEFORE;
-                    }
-                    else {
-                        return AFTER;
-                    }
-                }
-                else {
-                    return EQUAL;
-                }
-            }
-
-        }
-    }
-
-    private static class OrderLineObj {
-
-        private double distance;
-        private double classVal;
-
-        private OrderLineObj(double distance, double classVal) {
-            this.distance = distance;
-            this.classVal = classVal;
-        }
-    }
 
     // /**
     // *
